Add basic input functionality
Input is handled by a combination of standard C++ input files and
command line arguments, in the style of the python argparse library.
Specifically, scientific input should go through input files, as they
help keeping a permanent record of calculations and significantly
simplify repeats and restarts. Modifications to the code's overall
behaviour such as verbose mode, timing etc. should be passed through
command line arguments.

The first command line argument to be implemented is `-input`, which
passes to the code the path to a file from which to read the scientific
input parameters. Said parameters are stored in a struct and can
therefore be made available to any part of the code by simply including
the corresponding header file.