Demonstrate global struct usage for variable passing
Variables are to be declared and ordered in structs in vars.h,
which is included in each source file via the common.h header.
In addition, in order for each source file to be made aware of the
instance of a given global struct, an extern statement is to be
included in common.h.

The global structs are then initialised outside of the main function
in cmap.cpp and can be accessed and edited in every source file which
inlcudes common.h. This considerably simplifies variable passing into
and out of functions.