Implements most of Spin-Boson functionality
This includes initial condition sampling: For the
nuclear degrees of freedom from the Wigner transform
of the thermal, equilibrium of the uncoupled bath.
For the mapping variables from either the phi function
or its square.

Adds initialisation of the Spin-Boson parameters and
arrays to store potential and forces. Also implements
the Spin-Boson potential and force calculation, which
includes a shift of the potential and force matrix traces
to the state independent scalars.