feat: add Velocity-Verlet time integration
This commit adds the "simplest" integrator for the equations of
motion: the VelocityVerlet algorithm. Note that this algorithm
is notoriously sensitive to the timestep duration and becomes
unstable for too large a dt. This makes VV integration rather
inefficient. Explicit diagonalization should be used wherever
possible.

This commit also implements population autocorrelation functions
in addition to the Pauli matrix ones already in place.

Finally, the Spin-Boson input is moved to the new tests/ directory,
where example files and unit tests are to be kept from now on.