feat: add explicit diagonlization algorithm for eqs. of motion
The algorithms available to integrate the equations of motion are
now: Velocity Verlet and Explicit Diagonalization. The algorithm
can now be specified in the input file, using the alg string
keyword ["vverlet", "diag"]. The linear algebra components of the
Explicit Diagonalization algorithm are handled by the Eigen3
library.

The Explicit Diagonalization algorithm performs slightly slower
than Velocity Verlet but is expected to be more accurate for a lot
of systems and is not nearly as sensitive to the choice of timestep.
For systems consisting of many electronic states, Explicit
Diagonalization can scale very poorly, since each timestep involved
diagonalizing a (states,states) matrix twice.